Revamps chapters page UI and improves pagination
Redesigns the chapters listing page with a new layout, modern styles, and improved responsiveness. Simplifies logic for pagination and chapter display, removing the search feature and related state management. Adds new color and gap utilities to Tailwind config. Updates dependencies to include react-query.

Aims to enhance user experience and align with updated design guidelines.
