+++ package.json
@@ -33,6 +33,7 @@
     "react-hook-form": "^7.51.4",
     "react-icons": "^5.5.0",
     "react-lazy-load-image-component": "^1.5.6",
+    "react-query": "^3.39.3",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
