+++ src/pages/Chapters/Old.jsx
@@ -1,0 +1,258 @@
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useState } from "react";
+import { Helmet } from "react-helmet-async";
+import { useQuery } from "@tanstack/react-query";
+
+import Header from "../../components/Header";
+import searchIcon from "../../images/chapters/search-icon.png";
+import Footer from "../../components/Footer";
+import chaptersHero from "../../images/chapters/chapters-hero.png";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {
+  faAngleLeft,
+  faAngleRight,
+  faTriangleExclamation,
+} from "@fortawesome/free-solid-svg-icons";
+import { Link } from "react-router-dom";
+
+import { getChapters } from "../../services";
+import { apiConstants, paths } from "../../utils";
+import chapterImage from "../../images/chapters/sca-chapters-img.png";
+import JoinUs from "../../components/JoinUs";
+import ChaptersCard from "../../components/Chapters";
+import * as components from "../../components";
+
+const Chapters = () => {
+  const [page, setPage] = useState(1);
+  const [fullChapters, setFullChapters] = useState([]);
+  const [chapters, setChapters] = useState([]);
+  const [searchValue, setSearchValue] = useState("");
+  const [searchNotFound, setSearchNotFound] = useState(false);
+
+  useQuery(["chapters-full"], () => getChapters(1, 1000), {
+    onSuccess: (data) => {
+      setFullChapters(data.data);
+    },
+    enabled: true, // always run once on mount
+  });
+
+  const { data, isError, isLoading } = useQuery({
+    queryKey: [apiConstants.chapters, page],
+    queryFn: () => getChapters(page),
+    keepPreviousData: true,
+    onSuccess: (data) => {
+      setChapters(data.data);
+    },
+  });
+
+  const handleChange = (e) => {
+    if (!e.target.value) {
+      setChapters(data.data);
+    }
+    setSearchValue(e.target.value);
+  };
+
+  const handleSearch = () => {
+    if (!searchValue) {
+      setSearchNotFound(false);
+      setChapters(data.data);
+    }
+
+    const filteredResult = fullChapters.filter((chapter, index) => {
+      return (
+        chapter?.name.toLowerCase().includes(searchValue.toLowerCase()) ||
+        chapter?.city.toLowerCase().includes(searchValue.toLowerCase()) ||
+        chapter?.country.toLowerCase().includes(searchValue.toLowerCase())
+      );
+    });
+
+    if (filteredResult.length > 0) {
+      setSearchNotFound(false);
+      setChapters(filteredResult);
+    } else {
+      setSearchNotFound(true);
+      setChapters([]);
+    }
+  };
+
+  const nextPage = () => {
+    setPage(page + 1);
+  };
+
+  const prevPage = () => {
+    setPage(page - 1);
+  };
+
+  return (
+    <>
+      <Helmet>
+        <meta charSet="utf-8" />
+        <title>She Code Africa Chapters</title>
+        <meta
+          name="description"
+          content="You might want to consider starting a SheCodeAfrica Chapter in your campus or city. Not an expert in tech yet? Not a problem"
+        />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <meta property="og:title" content="She Code Africa Chapters" />
+        <meta
+          property="og:description"
+          content="You might want to consider starting a SheCodeAfrica Chapter in your campus or city. Not an expert in tech yet? Not a problem"
+        />
+        <meta name="twitter:title" content="She Code Africa Chapters" />
+        <meta
+          name="twitter:description"
+          content="You might want to consider starting a SheCodeAfrica Chapter in your campus or city. Not an expert in tech yet? Not a problem"
+        />
+      </Helmet>
+      <Header page={"chapters"} />
+      <main className=" text-secondary-main-black">
+        <section className=" bg-hero-bg-gradient pt-16 md:pt-24 lg:pt-28">
+          <div className="w-90 mx-auto min-h-[600px] flex flex-col justify-center 2md:justify-between 2md:flex-row md:items-center event-hero gap-8 py-12 2md:py-0 px-3 sm:px-0">
+            <div className="w-full max-w-[587px] ">
+              <h1 className="hero-heading capitalize font-bold text-[32px] md:text-[36px] 2md:text-[40px] text-center leading-[150%] mx-auto 2md:text-left">
+                Chapters in SCA
+              </h1>
+
+              <p className="m-0 mt-2 text-center md:w-[90%] text-base md:text-2xl mx-auto 2md:mx-0 2md:w-full leading-[30px] text-seal-brown 2md:text-justify font-normal">
+                Join 25,000+ African tech women on our platform and build your
+                successful career in the African tech industry. Explore our 43+
+                local chapters in 43 cities and campuses across Africa,
+                including Nigeria, Ghana, Kenya, Rwanda, and more.
+              </p>
+            </div>
+
+            <div className="w-full 2md:w-1/2">
+              <figure className="m-0 p-0 w-full max-w-[550.01px] h-[476.75px] mx-auto ">
+                <img
+                  src={chaptersHero}
+                  alt="ghana event"
+                  className="w-full h-full object-contain"
+                />
+              </figure>
+            </div>
+          </div>
+        </section>
+
+        <section className="border-2 border-gamboge pt-8 pr-8 pb-11 pl-9 mt-40 max-w-[820px] w-[90%] mx-auto rounded-[34.6111px] bg-floral-white flex items-center gap-8">
+          <FontAwesomeIcon
+            icon={faTriangleExclamation}
+            className="text-gamboge w-[52px] h-[45px]"
+          />
+
+          <p className="text-seal-brown font-medium">
+            She Code Africa(SCA) Chapters are independent chapters of the SCA
+            community and engagements do not speak directly on behalf of the
+            organisation except where explicitly stated.
+          </p>
+        </section>
+
+        <section className="mt-36 py-1">
+          <h2 className="text-3xl font-bold mb-0 lg:my-18 lg:text-[40px] text-center capitalize">
+            Start a chapter today
+          </h2>
+
+          <p className="w-full max-w-[832px] mx-auto mt-8 text-center text-lg">
+            Be a part of a community that celebrates diversity and empowers
+            women to thrive in tech. Connect with like-minded individuals, gain
+            access to resources and events, and help shape the future of the
+            industry.
+          </p>
+
+          <div className="flex items-center justify-center gap-8 mt-8">
+            <Link
+              to={paths.leadChapter}
+              className="bg-primary-main-pink px-8 py-[18px] mt-3 rounded-[30px] text-white text-base"
+            >
+              Lead a chapter
+            </Link>
+          </div>
+        </section>
+
+        <section className="bg-primary-main-pink py-[110px] md:px-24 mt-[154px]">
+          <h1 className="hero-heading capitalize font-extrabold text-[32px] md:text-[36px] 2md:text-[40px] text-center leading-[150%] mx-auto text-white">
+            Join a chapter
+          </h1>
+
+          <div className=" mt-14 flex flex-col sm:flex-row items-center justify-center gap-8 w-[90%] md:w-full mx-auto">
+            <div className="max-w-[310px] h-[54px] w-full py-3 px-2 bg-[#F7F7F7] text-xs rounded-[30px] overflow-hidden flex items-center pl-5">
+              <img src={searchIcon} alt="search" className="w-7" />
+              <input
+                className=" bg-transparent block focus:outline-none placeholder:text-veryLightGrey placeholder:text-base w-full h-full px-3"
+                type="search"
+                placeholder="Search"
+                value={searchValue}
+                // onChange={(e) => setSearchValue(e.target.value)}
+                onChange={handleChange}
+              />
+            </div>
+
+            <button
+              onClick={handleSearch}
+              className="bg-[#FDC0E3] px-8 py-4 inline-block mt-3 rounded-full text-[#434343] focus-visible:ring-1 focus-visible:ring-primary-main-pink"
+            >
+              Search
+            </button>
+          </div>
+
+          {searchNotFound && (
+            <h1 className="text-white capitalize text-center text-xl mt-11">
+              chapter not found.
+            </h1>
+          )}
+
+          {isError ? <components.Error /> : null}
+
+          {isLoading ? (
+            <div className="mt-[77px]">
+              <components.Loading />
+            </div>
+          ) : (
+            <>
+              <section className="grid grid-cols-1 sm:grid-cols-2 2md:grid-cols-4  mt-[77px] gap-8 w-[70%] mx-auto sm:w-[90%] 2md:w-full">
+                {chapters.length > 0 &&
+                  chapters.map((chapter, index) => {
+                    return (
+                      <div className="" key={index}>
+                        <ChaptersCard
+                          chapterImage={chapterImage}
+                          chapter={chapter}
+                        />
+                      </div>
+                    );
+                  })}
+              </section>
+            </>
+          )}
+
+          {!isLoading && (
+            <>
+              {data.data.length > 0 ? (
+                <div className="flex justify-center gap-7 mt-[57px]">
+                  <button
+                    className="bg-community-pink-bg border-0 w-[68px] h-[68px] overflow-hidden rounded-full disabled:bg-gray-400 disabled:text-white"
+                    onClick={prevPage}
+                    disabled={page === 1}
+                  >
+                    <FontAwesomeIcon icon={faAngleLeft} className="text-3xl" />
+                  </button>
+                  <button
+                    className="bg-community-pink-bg border-0 w-[68px] h-[68px] overflow-hidden rounded-full disabled:bg-gray-400 disabled:text-white"
+                    onClick={nextPage}
+                    disabled={data.totalPages === page}
+                  >
+                    <FontAwesomeIcon icon={faAngleRight} className="text-3xl" />
+                  </button>
+                </div>
+              ) : null}
+            </>
+          )}
+        </section>
+
+        <JoinUs />
+      </main>
+      <Footer />
+    </>
+  );
+};
+
+export default Chapters;
