+++ tailwind.config.js
@@ -39,6 +39,7 @@ module.exports = {
         "SCA-Blush": "#FFB8E0",
         "SCA-Citrine": "#FFF88F",
         "SCA-Forest": "#094A32",
+        "SCA-Apricote": "#FF8A3F",
         
       },
       fontSize: {
@@ -61,6 +62,9 @@ module.exports = {
         BebasNeue: ["Bebas Neue", "sans-serif"],
         figtree: ["Figtree", "sans-serif"],
       },
+      gap: {
+        13: "3.125rem"
+      }
     },
   },
   plugins: [],
